+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 #### Bug Fixes
 
+* Fix SQL queries for non-unique layer names
 * Respect webserver threads configuration (broken since 0.9.0)
 
 #### Breaking changes